requirements.txt
requests
tabulate
pygments
pydantic[dotenv]
pyyaml
python-jose[cryptography]
